Don't store certificate type strings on the heap
Hardcode the certificate type strings into GetCertType() (putting them
on the stack) and use an if-else to determine the appropriate value of
CertType to return. The performance gains of this change are negligable.
It might be ever so slightly faster (1, maybe 2 ms), and ever so
slightly more memory efficient. However, the primary advantage is that
it is less error prone as it doesn't involve keeping the values of the
CertType enumeration in sync with the indices of the array storing the
certificate type strings.

Additionally, it makes it easier to re-order the certificate type
strings so that more common values can be tested first, providing a
potential speed boost.